latest Lombok and TG api lib

+++ telegram-bot/src/main/java/com/instagirls/model/TelegramUser.java
@@ -17,7 +17,7 @@ public class TelegramUser {
     private UUID uuid;
 
     @Column(nullable = false, unique = true)
-    private Integer telegramId;
+    private Long telegramId;
 
     @Column
     private String firstName;
